+++ NEWS.md
@@ -2,7 +2,11 @@
 
 * Added `find_sequence()`, `set_contents()`.
 * Add `DEFINITION` tag to allow `\begin` etc. to be temporarily
-ignored.
+ignored.  This allows parseLatex to parse `\newenvironment` and
+related macros that create definitions.
+* `parseLatex()` now stops parsing after `\end{document}` just
+as LaTeX does.  The `get_leftovers()` function can retrieve
+the unparsed text.
 
 # parseLatex 0.1.0